# apps/books-service/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing books in Shelfshare.",
        "title": "Shelfshare Books API",
        "contact": {
            "name": "Sina Niyavarzi",
            "email": "sinaniya@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/books": {
            "get": {
                "description": "Get all books",
                "produces": ["application/json"],
                "tags": ["books"],
                "summary": "List books",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at_desc",
                            "created_at_asc",
                            "title_asc",
                            "title_desc",
                            "published_at_desc",
                            "published_at_asc"
                        ],
                        "type": "string",
                        "description": "Sort field and direction",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Full-text search on title and description",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by author ID (UUID)",
                        "name": "author_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2015-01-01",
                        "description": "Filter: published_at \u003e= YYYY-MM-DD",
                        "name": "published_after",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2020-12-31",
                        "description": "Filter: published_at \u003c= YYYY-MM-DD",
                        "name": "published_before",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ListBooksResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new book with title, author, description and optional published date",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["books"],
                "summary": "Create a book",
                "parameters": [
                    {
                        "description": "Book to create",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.BookResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Login required",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "description": "Get a single book by its UUID",
                "produces": ["application/json"],
                "tags": ["books"],
                "summary": "Get a book by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BookResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a book by its UUID",
                "produces": ["application/json"],
                "tags": ["books"],
                "summary": "Delete a book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Login required",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially update a book by its UUID",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["books"],
                "summary": "Update a book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BookResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID or payload",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Login required",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Book not found",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.AuthorSummary": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handler.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/handler.AuthorSummary"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-11-24"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string",
                    "example": "2025-11-24"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-11-24"
                }
            }
        },
        "handler.BookResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handler.Book"
                }
            }
        },
...
        "handler.CreateBookRequest": {
            "type": "object",
            "required": ["author_id", "title"],
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string",
                    "example": "2025-11-24"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handler.ListBooksResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.Book"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/handler.Pagination"
                }
            }
        },
        "handler.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "minimum": 1
                },
                "page_size": {
                    "type": "integer",
                    "minimum": 1
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "handler.UpdateBookRequest": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "published_at": {
                    "type": "string",
                    "example": "2025-11-24"
                },
                "title": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "validation.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validation.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "validation.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "rule": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Supply the auth-service JWT as: Bearer \u003ctoken\u003e. Required for all mutations.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}